[BF] 挤掉之后的闪退问题      客户列表点击顺序错乱问题      token过期，关闭所有界面，退回到登陆界面      适配退货率按月表示      报表接口收藏数字段改成week_collection，退货率改成refund      登录跳转错误问题      柱状图适配0.01 ～ ∞      红米客户详情无法滑动问题 [WN] 修改多渠道和release打包，与混淆文件

+++ app/src/main/java/com/soubu/goldensteward/GoldenStewardApplication.java
@@ -30,10 +30,6 @@ import org.greenrobot.eventbus.EventBus;
 
 import java.util.List;
 
-import static android.R.id.list;
-import static com.baidu.location.h.j.P;
-import static com.baidu.location.h.j.u;
-
 /**
  * Created by dingsigang on 16-10-18.
  */
@@ -136,7 +132,7 @@ public class GoldenStewardApplication extends Application implements Application
 
 
     public void saveUserInfo(UserServerParams params) {
-        if(dao == null){
+        if (dao == null) {
             dao = DBHelper.getInstance(sInstance).getUserDao();
         }
         List<User> list = dao.queryBuilder().where(UserDao.Properties.Phone.eq(params.getPhone())).list();
@@ -173,13 +169,15 @@ public class GoldenStewardApplication extends Application implements Application
         setPhone(params.getPhone());
     }
 
-    public void clearUser(){
-        dao.delete(user);
+    public void clearUser() {
+        if (dao != null) {
+            dao.delete(user);
+        }
     }
 
-    public boolean initUser(){
+    public boolean initUser() {
         dao = DBHelper.getInstance(sInstance).getUserDao();
-        if(!TextUtils.isEmpty(getPhone())){
+        if (!TextUtils.isEmpty(getPhone())) {
             List<User> list = dao.queryBuilder().where(UserDao.Properties.Phone.eq(getPhone())).list();
             if (list.size() > 0) {
                 user = list.get(0);
